Widen the HEIF stride check past i32
The plane guard compared stride against width * 4 in i32, which wraps for
a width past ~536 million and would let an undersized stride through into
the per-row read. The multiply now widens to i64. libheif caps decoded
dimensions far below that, so every reachable image behaves exactly as
before; the check is simply correct on its own terms now.

+++ src/image/fallback.rs
@@ -426,7 +426,7 @@ fn decode_heif_primary_image(
     let mut stride: c_int = 0;
     let plane =
         unsafe { heif_image_get_plane_readonly(image, HEIF_CHANNEL_INTERLEAVED, &raw mut stride) };
-    if plane.is_null() || width <= 0 || height <= 0 || stride < width * 4 {
+    if plane.is_null() || width <= 0 || height <= 0 || i64::from(stride) < i64::from(width) * 4 {
         unsafe { heif_image_release(image) };
         return Err(uncoded_error("HEIF image plane unavailable"));
     }
